test_email_service/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
behave==1.2.6
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.4
cucumber-tag-expressions==4.1.0
dnspython==2.2.1
email-validator==1.3.0
Faker==14.2.1
h11==0.13.0
idna==3.3
outcome==1.2.0
parse==1.19.0
parse-type==0.6.0
pycparser==2.21
pydantic==1.10.2
pydantic-factories==1.6.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
splinter==0.18.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
win-unicode-console==0.5
wsproto==1.2.0
xeger==0.3.5
